requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.1.0
azure-ai-inference==1.0.0b9
azure-ai-ml==1.25.0
azure-common==1.1.28
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-identity==1.20.0
azure-mgmt-core==1.5.0
azure-monitor-opentelemetry==1.6.5
azure-monitor-opentelemetry-exporter==1.0.0b33
azure-storage-blob==12.24.1
azure-storage-file-datalake==12.18.1
azure-storage-file-share==12.20.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.10
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.3.2
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docx2pdf==0.1.8
durationpy==0.9
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
fixedint==0.1.6
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-ollama==0.2.3
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langchain-unstructured==0.1.6
langdetect==1.0.9
langsmith==0.3.8
libmagic==1.0
libreserver==0.1.1
lxml==5.3.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.7
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.63.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-dbapi==0.51b0
opentelemetry-instrumentation-django==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-flask==0.51b0
opentelemetry-instrumentation-psycopg2==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-urllib==0.51b0
opentelemetry-instrumentation-urllib3==0.51b0
opentelemetry-instrumentation-wsgi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandoc==2.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pi_heif==0.21.0
pikepdf==9.7.0
pillow==11.1.0
plumbum==1.9.0
ply==3.11
poppler-utils==0.1.0
portalocker==2.10.1
posthog==3.14.2
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==5.9.8
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydash==8.0.5
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMuPDF==1.25.3
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.3.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytweening==1.2.0
pytz==2025.1
pywin32==309
PyYAML==6.0.2
ragas==0.2.14
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.0
rsa==4.9
safetensors==0.5.2
scipy==1.15.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
strictyaml==1.7.3
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.23
unstructured-client==0.30.4
unstructured-inference==0.8.10
unstructured.pytesseract==0.3.15
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
